Composer: raise the minimum supported PHPCS version to 3.8.0
... to prevent end-users from running into trouble with the name change.

The files in the Composer `vendor/bin` will only be replaced when the `replace...` directive is found and that is only available in the 3.8.0 tag.
When the files in the Composer `vendor/bin` are not replaced, they will continue to point to the `vendor/squizlabs/php_codesniffer` directory which will no longer exist, leading to fatal "File not found" errors for end-users trying to run PHPCS/PHPCBF.

Includes updating references to the PHPCS version whenever relevant throughout the codebase.